Swap to data actions menu on data page (#80)
Swap to data actions menu on data page and tweak tab view nav/toolbar styling.

Menu name is a bit awkward and needs discussion in the New Year. Still, an improvement on the buttons.

We discussed using a burger menu for this but I'd like to do better. I think ideally we'd have a heading for the table of samples and this would be a triple dot menu alongside the heading but that needs more discussion.