Make the frontend aware of the magnetometer.
The magnetometer data is currently just logged to the terminal,
but is ready to be accessed in the same way the accelerometer
service.

I'm not sure why, but it seems that the microbit is
not correctly sending its UART initialisation sequence,
hence in the UI an alert about outdated firmware pops up.